# docs/swagger.yaml
definitions:
...
  github_com_michaelwp_student_attendance_internal_models.Teacher:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      phone:
        type: string
      teacher_id:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /teachers:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of all teachers
      parameters:
      - default: 10
        description: Number of teachers to return (max 100)
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of teachers to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of teachers retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get all teachers
      tags:
      - Teachers
    post:
      consumes:
      - application/json
      description: Create a new teacher in the system
      parameters:
      - description: Teacher data
        in: body
        name: teacher
        required: true
        schema:
          $ref: '#/definitions/github_com_michaelwp_student_attendance_internal_models.Teacher'
      produces:
      - application/json
      responses:
        "201":
          description: Teacher created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new teacher
      tags:
      - Teachers
  /teachers/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a teacher from the system
      parameters:
      - description: Teacher database ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Teacher deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid teacher ID
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Delete teacher
      tags:
      - Teachers
    get:
      consumes:
      - application/json
      description: Retrieve a specific teacher by their database ID
      parameters:
      - description: Teacher database ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Teacher retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid teacher ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Teacher not found
          schema:
            additionalProperties: true
            type: object
      summary: Get teacher by ID
      tags:
      - Teachers
    put:
      consumes:
      - application/json
      description: Update a teacher's information
      parameters:
      - description: Teacher database ID
        in: path
        name: id
        required: true
        type: integer
      - description: Teacher data
        in: body
        name: teacher
        required: true
        schema:
          $ref: '#/definitions/github_com_michaelwp_student_attendance_internal_models.Teacher'
      produces:
      - application/json
      responses:
        "200":
          description: Teacher updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Update teacher
      tags:
      - Teachers
  /teachers/teacher-id/{teacherId}:
    get:
      consumes:
      - application/json
      description: Retrieve a specific teacher by their teacher ID
      parameters:
      - description: Teacher ID (not database ID)
        in: path
        name: teacherId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Teacher retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Teacher ID is required
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Teacher not found
          schema:
            additionalProperties: true
            type: object
      summary: Get teacher by teacher ID
      tags:
      - Teachers
swagger: "2.0"